+++ Model/Method/Facilypay/AbstractFacilypay.php
@@ -24,10 +24,6 @@ use Magento\Framework\Exception\LocalizedException;
 
 class AbstractFacilypay extends AbstractMethodAPI
 {
-    /**
-     * @var int 48H
-     */
-    public $overridePendingTimeout = 2880;
 
     /**
      * Validate payment method information object